Move yaml options into their own base config class
