+++ keyconnect-cli/src/main/java/app/keyconnect/cli/config/BaseClientConfig.java
@@ -2,12 +2,10 @@ package app.keyconnect.cli.config;
 
 import app.keyconnect.api.ApiClient;
 import app.keyconnect.api.client.DefaultApi;
-import app.keyconnect.cli.utils.ConsoleMode;
-import app.keyconnect.cli.utils.ConsoleUtil;
 import org.jetbrains.annotations.NotNull;
 import picocli.CommandLine.Option;
 
-public class BaseClientConfig {
+public class BaseClientConfig extends BaseConfig {
 
   @Option(
       names = {"-s", "--server"},
@@ -22,26 +20,6 @@ public class BaseClientConfig {
       defaultValue = "false"
   )
   private boolean apiDebug;
-
-  @Option(
-      names = {"--print-json", "--json"},
-      description = "Print output in JSON format",
-      defaultValue = "false"
-  )
-  private void setJson(boolean value) {
-    if (value)
-      ConsoleUtil.setMode(ConsoleMode.Json);
-  }
-
-  @Option(
-      names = {"--print-yaml", "--yaml"},
-      description = "Print output in YAML format",
-      defaultValue = "false"
-  )
-  private void setYaml(boolean value) {
-    if (value)
-      ConsoleUtil.setMode(ConsoleMode.Yaml);
-  }
 
   @NotNull
   protected DefaultApi buildApiClient() {
